+++ apig-20240327/alibabacloud_apig20240327/models/_ai_service_config.py
@@ -2,7 +2,7 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
-from typing import List
+from typing import List, Dict
 
 from alibabacloud_apig20240327 import models as main_models
 from darabonba.model import DaraModel
@@ -14,25 +14,35 @@ class AiServiceConfig(DaraModel):
         address: str = None,
         api_keys: List[str] = None,
         bedrock_service_config: main_models.AiServiceConfigBedrockServiceConfig = None,
+        compatible_protocols: List[str] = None,
+        default_model_name: str = None,
         enable_health_check: bool = None,
+        enable_outlier_detection: bool = None,
         pai_easservice_config: main_models.AiServiceConfigPaiEASServiceConfig = None,
         protocols: List[str] = None,
         provider: str = None,
+        vertex_service_config: main_models.AiServiceConfigVertexServiceConfig = None,
     ):
         self.api_key_generate_mode = api_key_generate_mode
         self.address = address
         self.api_keys = api_keys
         self.bedrock_service_config = bedrock_service_config
+        self.compatible_protocols = compatible_protocols
+        self.default_model_name = default_model_name
         self.enable_health_check = enable_health_check
+        self.enable_outlier_detection = enable_outlier_detection
         self.pai_easservice_config = pai_easservice_config
         self.protocols = protocols
         self.provider = provider
+        self.vertex_service_config = vertex_service_config
 
     def validate(self):
         if self.bedrock_service_config:
             self.bedrock_service_config.validate()
         if self.pai_easservice_config:
             self.pai_easservice_config.validate()
+        if self.vertex_service_config:
+            self.vertex_service_config.validate()
 
     def to_map(self):
         result = dict()
@@ -51,8 +61,17 @@ class AiServiceConfig(DaraModel):
         if self.bedrock_service_config is not None:
             result['bedrockServiceConfig'] = self.bedrock_service_config.to_map()
 
+        if self.compatible_protocols is not None:
+            result['compatibleProtocols'] = self.compatible_protocols
+
+        if self.default_model_name is not None:
+            result['defaultModelName'] = self.default_model_name
+
         if self.enable_health_check is not None:
             result['enableHealthCheck'] = self.enable_health_check
+
+        if self.enable_outlier_detection is not None:
+            result['enableOutlierDetection'] = self.enable_outlier_detection
 
         if self.pai_easservice_config is not None:
             result['paiEASServiceConfig'] = self.pai_easservice_config.to_map()
@@ -62,6 +81,9 @@ class AiServiceConfig(DaraModel):
 
         if self.provider is not None:
             result['provider'] = self.provider
+
+        if self.vertex_service_config is not None:
+            result['vertexServiceConfig'] = self.vertex_service_config.to_map()
 
         return result
 
@@ -80,8 +102,17 @@ class AiServiceConfig(DaraModel):
             temp_model = main_models.AiServiceConfigBedrockServiceConfig()
             self.bedrock_service_config = temp_model.from_map(m.get('bedrockServiceConfig'))
 
+        if m.get('compatibleProtocols') is not None:
+            self.compatible_protocols = m.get('compatibleProtocols')
+
+        if m.get('defaultModelName') is not None:
+            self.default_model_name = m.get('defaultModelName')
+
         if m.get('enableHealthCheck') is not None:
             self.enable_health_check = m.get('enableHealthCheck')
+
+        if m.get('enableOutlierDetection') is not None:
+            self.enable_outlier_detection = m.get('enableOutlierDetection')
 
         if m.get('paiEASServiceConfig') is not None:
             temp_model = main_models.AiServiceConfigPaiEASServiceConfig()
@@ -92,6 +123,77 @@ class AiServiceConfig(DaraModel):
 
         if m.get('provider') is not None:
             self.provider = m.get('provider')
+
+        if m.get('vertexServiceConfig') is not None:
+            temp_model = main_models.AiServiceConfigVertexServiceConfig()
+            self.vertex_service_config = temp_model.from_map(m.get('vertexServiceConfig'))
+
+        return self
+
+class AiServiceConfigVertexServiceConfig(DaraModel):
+    def __init__(
+        self,
+        gemini_safety_setting: Dict[str, str] = None,
+        vertex_auth_key: str = None,
+        vertex_auth_service_name: str = None,
+        vertex_project_id: str = None,
+        vertex_region: str = None,
+        vertex_token_refresh_ahead: int = None,
+    ):
+        self.gemini_safety_setting = gemini_safety_setting
+        self.vertex_auth_key = vertex_auth_key
+        self.vertex_auth_service_name = vertex_auth_service_name
+        self.vertex_project_id = vertex_project_id
+        self.vertex_region = vertex_region
+        self.vertex_token_refresh_ahead = vertex_token_refresh_ahead
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.gemini_safety_setting is not None:
+            result['geminiSafetySetting'] = self.gemini_safety_setting
+
+        if self.vertex_auth_key is not None:
+            result['vertexAuthKey'] = self.vertex_auth_key
+
+        if self.vertex_auth_service_name is not None:
+            result['vertexAuthServiceName'] = self.vertex_auth_service_name
+
+        if self.vertex_project_id is not None:
+            result['vertexProjectId'] = self.vertex_project_id
+
+        if self.vertex_region is not None:
+            result['vertexRegion'] = self.vertex_region
+
+        if self.vertex_token_refresh_ahead is not None:
+            result['vertexTokenRefreshAhead'] = self.vertex_token_refresh_ahead
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('geminiSafetySetting') is not None:
+            self.gemini_safety_setting = m.get('geminiSafetySetting')
+
+        if m.get('vertexAuthKey') is not None:
+            self.vertex_auth_key = m.get('vertexAuthKey')
+
+        if m.get('vertexAuthServiceName') is not None:
+            self.vertex_auth_service_name = m.get('vertexAuthServiceName')
+
+        if m.get('vertexProjectId') is not None:
+            self.vertex_project_id = m.get('vertexProjectId')
+
+        if m.get('vertexRegion') is not None:
+            self.vertex_region = m.get('vertexRegion')
+
+        if m.get('vertexTokenRefreshAhead') is not None:
+            self.vertex_token_refresh_ahead = m.get('vertexTokenRefreshAhead')
 
         return self